requirements.txt
# Data Processing & Analysis
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pymupdf==1.24.5

# Streamlit and UI
streamlit==1.36.0
streamlit-option-menu==0.3.13
streamlit-aggrid==0.3.4.post3

# Visualization
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2
altair==5.3.0

# Workflow Orchestration
prefect==2.19.9
prefect-client==2.19.9

# Google Cloud (for ADK)
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# HTTP Requests & APIs
requests==2.32.3
httpx==0.27.0
aiohttp==3.9.5

# Configuration & Environment
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Logging & Monitoring
loguru==0.7.2
rich==13.7.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
pytest-cov==5.0.0

# Data Validation
marshmallow==3.21.3
cerberus==1.3.5

# File I/O
openpyxl==3.1.5
xlsxwriter==3.2.0

# Date/Time handling
python-dateutil==2.9.0.post0

# String processing
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1

# Async support
asyncio-throttle==1.0.2

# Performance monitoring
psutil==6.0.0

# Security
cryptography==42.0.8

# Windows compatibility
pywin32==306; sys_platform == "win32"